requirements.txt
streamlit 
pypdf2
langchain
langchain_core
langchain_community
langchain-groq
python-dotenv
faiss-cpu
huggingface_hub
openai
InstructorEmbedding
sentence-transformers==2.2.2
torch